Stricter serialization of annotation namespaces
If the user already has namespace a = 'something else', the code for
a:documentation (## comments) must include an xmlns:a='...' to
point it at http://relaxng.org/ns/compatibility/annotations/1.0
despite that redefinition

It should be an an error for an annotation element to use an undefined
namespace prefix (otherwise we produce invalid XML)

annotations.rnc should define the a: prefix it's using,
trang requires this to parse the file at all, and I can find no
justification in the compact syntax spec for pre-defining
namespace a = 'http://relaxng.org/ns/compatibility/annotations/1.0'.

But leave the default in anyway, since it should be harmless
now that it's overridden by any conflicting definition;
removing it would break backward-compatibility.